Bug with multiview_panels
When launching multiview panels plugin on the world
and the StereoRL single camera results in different
windows sizes. The plugin should be able to control
the window size but something inside AMBF changes
changes internal the size of the window.

To reproduce select this commit, build and run ambf
with:

```
ambf_simulator --launch_file launch.yaml -l 0,2,5,6
```